// packages/bitcoin/src/gql/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  AssetV0Args: any;
  /**
   * A datetime with timezone offset.
   *
   * The input is a string in RFC3339 format, e.g. "2022-01-12T04:00:19.12345Z"
   * or "2022-01-12T04:00:19+03:00". The output is also a string in RFC3339
   * format, but it is always normalized to the UTC (Z) offset, e.g.
   * "2022-01-12T04:00:19.12345Z".
   */
  DateTime: any;
  /** Decimal (fixed-point) */
  Decimal: any;
  IntegerString: any;
  /** A scalar that can represent any JSON value. */
  JSON: any;
  /**
   * A local datetime without timezone offset.
   *
   * The input/output is a string in ISO 8601 format without timezone, including
   * subseconds. E.g. "2022-01-12T07:30:19.12345".
   */
  LocalDateTime: any;
};
// ...
export type Amount = {
  __typename?: 'Amount';
  /** @deprecated use `decimals` from CryptoAsset */
  scalingFactor?: Maybe<Scalars['Int']>;
  value: Scalars['String'];
};
// ...
export type FloorPrice = {
  __typename?: 'FloorPrice';
  marketplaceId: Scalars['String'];
  paymentToken: PaymentToken;
  value?: Maybe<Scalars['String']>;
  valueUsdCents?: Maybe<Scalars['String']>;
};
// ...
export type Marketplace = {
  __typename?: 'Marketplace';
  collectionUrl: Scalars['String'];
  logoUrl?: Maybe<Scalars['String']>;
  marketplaceCollectionId: Scalars['String'];
  marketplaceId: Scalars['String'];
  marketplaceName: Scalars['String'];
  nftUrl: Scalars['String'];
  verified?: Maybe<Scalars['Boolean']>;
};
// ...
export type MediaV2 = {
  __typename?: 'MediaV2';
  /** contains blurhash value of the media (if media type supports it) */
  blurHash?: Maybe<Scalars['String']>;
  /**
   * represents actual content type by peeking into content referenced by media URL
   * format as described in: <https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types>
   * i.e. image/png if url points to PNG image
   */
  contentType?: Maybe<Scalars['String']>;
  /**
   * "type" field describes "source" of media URL from metadata.json
   * if URL is taken from "image" or "image_url" fields - type will have image value
   * if URL is taken from "animated_url" field - type will be animated
   */
  type: Scalars['String'];
  url: Scalars['String'];
};
// ...
export type NftAttribute = {
  __typename?: 'NFTAttribute';
  displayType?: Maybe<Scalars['String']>;
  traitType: Scalars['String'];
  value: Scalars['JSON'];
};

export type NftCollectionV2 = {
  __typename?: 'NFTCollectionV2';
  address: Scalars['String'];
  media?: Maybe<MediaV2>;
  name: Scalars['String'];
  symbol: Scalars['String'];
};

export type NftCollectionV3 = {
  __typename?: 'NFTCollectionV3';
  address: Scalars['String'];
  collectionItemsAmount?: Maybe<Scalars['String']>;
  collectionItemsOwnersAmount?: Maybe<Scalars['String']>;
  floorPrices?: Maybe<Array<FloorPrice>>;
  marketplaces?: Maybe<Array<Marketplace>>;
  media?: Maybe<MediaV2>;
  name?: Maybe<Scalars['String']>;
  symbol: Scalars['String'];
};

export type NftLastSale = {
  __typename?: 'NFTLastSale';
  cryptoPrice: Amount;
  fiatPrice: Amount;
  quantity: Amount;
};

export type NftLastSaleV2 = {
  __typename?: 'NFTLastSaleV2';
  cryptoPrice?: Maybe<Amount>;
  fiatPrice?: Maybe<Amount>;
  quantity: Amount;
};
// ...
export type NfTv3 = {
  __typename?: 'NFTv3';
  attributes: Array<NftAttribute>;
  balance: Amount;
  /** Chain name */
  chain?: Maybe<Scalars['String']>;
  /** @deprecated NFTCollectionV2 is deprecated. Please move to the NFTCollectionV3 */
  collection?: Maybe<NftCollectionV2>;
  collectionV3?: Maybe<NftCollectionV3>;
  /** Crypto currency address on specific chain */
  contract?: Maybe<Scalars['String']>;
  contractType?: Maybe<Scalars['String']>;
  description?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  isNftSpam: Scalars['Boolean'];
  lastSale?: Maybe<NftLastSaleV2>;
  location?: Maybe<Scalars['String']>;
  media: Array<MediaV2>;
  name: Scalars['String'];
  owner: Scalars['String'];
  spamScore?: Maybe<Scalars['Float']>;
  symbol: Scalars['String'];
  tokenId?: Maybe<Scalars['String']>;
};